+++ app/models/webdav/segmentation.rb
@@ -4,8 +4,8 @@ module Webdav
   class Segmentation < Webdav::Client
     def initialize
       super(storage_url)
-      credentials(Rails.application.config_for('eurvalve')['seg_service']['user'],
-                  Rails.application.config_for('eurvalve')['seg_service']['password'])
+      credentials(Rails.application.config_for('eurvalve')['segmentation']['user'],
+                  Rails.application.config_for('eurvalve')['segmentation']['password'])
       self.verify_server = false
     end
 
@@ -21,18 +21,18 @@ module Webdav
       private
 
       def inputs_path
-        Rails.application.config_for('eurvalve')['seg_service']['inputs_path']
+        Rails.application.config_for('eurvalve')['segmentation']['inputs_path']
       end
 
       def outputs_path
-        Rails.application.config_for('eurvalve')['seg_service']['outputs_path']
+        Rails.application.config_for('eurvalve')['segmentation']['outputs_path']
       end
     end
 
     private
 
     def storage_url
-      Rails.application.config_for('eurvalve')['seg_service']['url']
+      Rails.application.config_for('eurvalve')['segmentation']['url']
     end
   end
 end