+++ app/models/flow.rb
@@ -99,8 +99,7 @@ class Flow
     RimrockStep.new('valve_sizing', 'eurvalve/mock-step', 'mock.sh.erb'),
     RimrockStep.new('0d_models',
                     'eurvalve/0dmodel', '0d_scenarios.sh.erb',
-                    [:parameter_optimization_result]),
-    CloudStep.new('sample_cloud_step', 'eurvalve/mock-step', 'cloud_mock.sh.erb')
+                    [:parameter_optimization_result])
   ].freeze
 
   steps_hsh = Hash[STEPS.map { |s| [s.name, s] }]