+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 - Added AVD/MVD ratio to patients' statistics (@Nuanda)
 
 ### Changed
+- Reintegration of segmentation service using File Store in place of OwnCloud (@jmeizner)
 
 ### Deprecated